requirements.txt
numpy>=1.20.0
pyusb>=1.2.1
pyserial>=3.5
matplotlib>=3.5.0
scipy>=1.7.0
sqlalchemy>=1.4.0
pyhackrf
pyrtlsdr>=0.2.91
Flask>=2.0.0
Flask-SocketIO>=5.0.0
ptyprocess>=0.7.0